requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
datasets==4.4.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.1.4
idna==3.11
jinja2==3.1.6
jiter==0.12.0
markupsafe==3.0.3
multidict==6.7.0
multiprocess==0.70.18
networkx==3.5
numpy==2.3.4
openai==2.8.0
packaging==25.0
pandas==2.3.3
propcache==0.4.1
pyarrow==22.0.0
pydantic==2.12.4
pydantic-core==2.41.5
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
